Add matplotlib intersphinx to the docs
This commit adds the docs configuration to leverage the intersphinx
plugin [1] to link to the matplotlib documentation natively inline in
docstrings like it were part of the same sphinx document. This then
updates the docstrings to use references to the actual matplotlib
objects instead of referring to the internal import paths for matplotlib
objects.

[1] https://www.sphinx-doc.org/en/master/usage/extensions/intersphinx.html