Add missing VERSION.txt file to the manifest
The version.txt file is necessary to build qiskit-experiments from
source since it's used by the setup.py to determing the version in the
package metadata. However this wasn't being included in the packaged
sdist because it was missing from the manifest. This causes builds from
sdist to fail because setup.py can't find the version file. This commit
fixes this oversight and adds the missing entry to the manifest so that
we're including all the necessary components in the sdist.