Configure docs publishing job for test branch
This commit tweaks the docs publishing job to publish the documentation
for the test branch to https://qiskit.org/documentation/experiments/test
on every merged commit to the test branch. When we release
qiskit-experiments the docs job from the main branch will be triggered
deleting these published docs from cloud object storage as the publish
job runs 'rclone sync' which deletes files in target not in source. This
should give us the desired state where the test docs only exist as long
as the test branch does prior to the first release of
qiskit-experiments.