Leverage concurrency groups to auto cancel jobs
This commit leverages a new beta feature in github actions, concurrency
groups [1], to enable job auto cancellation when there is a PR update.
Basically we define a unique concurrency group for each PR and limit the
the group to only run a single workflow at once and cancel any existing
jobs when a new one is launched. This should hopefully improve CI
throughput by not running duplicate jobs on PR updates.

[1] https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions#concurrency