Set minimum python version to 3.7 (#227)
This commit changes the minimum supported python version from 3.6 to
3.7. qiskit-experiments uses dataclasses internally which was a
python feature introduced in 3.7 and the code as packaged doesn't
actually work on 3.6. This hasn't been caught in CI because a test
dependency of qiskit-experiments is installing the dataclasses backport
package as a dependency which is masking this issue. While we can list
that package as a depedency of qiskit-experiments too to resolve this,
python 3.6 is deprecated in the upstream qiskit packages and the next
qiskit-terra minor version release (0.19.0) will be the last to support
it. At this point it just makes more sense to drop 3.6 support.

At the same time the minimum supported version of the qiskit-terra
package is bumped to the latest release 0.18.0 which is actually the
minimum version needed (we were installing from git in CI to workaround
this pre-0.18.0 release) and 0.17.0 was the release which deprecated
python 3.6 in qiskit which lines up well with dropping 3.6 support here.