Only scale y axis in curve_analysis if fit_data is present
In cases when there is no fit_data present trying to automatically scale
the y axis based on the y range of the fit data doesn't make sense. This
is causing a failure in CI now that mpl is being installed. This commit
fixes this by only running the scaling if there is fit_data.