Make matplotlib a requirement (#280)
This commit makes matplotlib a hard requirement for qiskit experiments.
Plotting is an integral component to analysis and almost every user will
need to install matplotlib to work with experiments. So instead of
having workarounds to keep it optional, this commit just adds it to the
requirements list and removes all the workarounds in place to do runtime
imports and detection of it missing.