Workaround reno issue
The current release of reno has a bug where it expects there to be a
master branch present when it needs to scan the git history to figure
out where tags are relative to commits. However the experiments repo
does not have a master branch (and never did). To workaround this
limitation there are 2 potential fixes. Either add a master branch
temporarily until the release (which is what qiskit-terra did) or
install reno from source which already has a fix [1] for this. This
commit opts for the latter because the qiskit-experiments repo has never
had a master branch so creating a throwaway branch for this seems
unecessary. This does mean people will need to install reno from source
if they run outside of tox.

[1] https://opendev.org/openstack/reno/commit/ed6bbae82e01edf781534d9e9cce21e0c55a49a6