Use latest reno release
In #294 we switched to using reno from source to workaround a bug in
reno that required a branch named master to exist for
When #294 was proposed the fix for this issue [1] was not included in
a release. But since #294 has merged a new reno release has been
published [2] and we can instead rely on that version instead of the
installing from source. This commit makes this change and switches
to using reno>=3.4.0 instead of reno from git.

[1] https://opendev.org/openstack/reno/commit/ed6bbae82e01edf781534d9e9cce21e0c55a49a6
[2] https://pypi.org/project/reno/3.4.0/