Add missing __init__.py to make qiskit_experiments.test importable (#395)
This commit fixes a packaging issue with the qiskit_experiments.test
package. It was missing it's __init__.py file meaning it wasn't a valid
python package and couldn't be imported when install from a a package.
This wasn't caught in CI because we have it locally and the path is
found in the git tree. This commit adds the missing file to fix this
and adds it to the documentation so there are docs available for the
packaged test module.

Fixes #393

Co-authored-by: Helena Zhang <Helena.Zhang@ibm.com>