Fix DB display for non-float values (#301)
* Fix DB display for non-finite floats

* Add safe JSON serialization of inf and NaN

* Fix some issues with safe float and recusion

* A few more fixes

* Add string conversion for result display

Add ability to display complex, list, and array results in result DB by converting to string. Currently this limits display to cases where the resulting string is < 60 characters.

* Add tests for display conversion

* Fix array2string conversion

* Remove special handling for non float values

Now only complex numbers are converted to strings, and any string values are uploaded to the database without checking their length (the service should do its own checking and truncation if required). For values that are not a float or string type, they are uploaded as a string of the class name "(cls)".

* Fixup tests for renamed method

* Remove unused import

* Include safe nan handling for chisq

Co-authored-by: Yael Ben-Haim <yaelbh@il.ibm.com>