Review comments, rework child_ids
Change child data to be stored as a ThreadSafeOrderedDict instead of ThreadSafeList, where the keys are the child experiment ids. This allows child data to also be looked up by experiment id as well as index. The child ids are now only copied to metadata when saving to DB, and popped back off metadata when loading.