Add StoreInitArgs mixin
This generalizes the Settings mixin to allow storing init args for classes with variadic args which will make life easier when subclassing Expeirment and Anaysis classes with trivial init methods since they can use `functools.wraps` and (*args, **kwargs)` for their parent class. The Settings mixin is now a subclass of this mixin which also adds the `settings` property and raises an exception if used in a class with variadic args in its init method.