Retrieve data when loading ExperimentData
Previously when loading an ExperimentData from the result database
the raw data was not loaded from the provider job database until the
`ExperimentData.data()` method was called. This changes the
`ExperimentData.load` method do this retrieval as part of loading
the data from the database service rather than doing it lazily only
when `data()` was called.