Fix Sphinx build on Python 3.9+ (#612)
On Python 3.9, Sphinx requires that forwards references (type hints with
their names in strings) resolve to actual names after the document is
parsed.  There is a cyclical dependency that prevents the name from
being imported properly, but we can using `typing.TYPE_CHECKING` to
limit the import to only analysis tools (like Sphinx here) that are able
to handle it.