fix experiment equiv check in unittest
In current equivalence checking we assume all args or kwargs value are not Experiment nor Analysis object. However if we need to write round-trip json test for composite experiment indeed args are these object and test fails. New code checks each element in the arg and apply special equivalence check if the element is Experiment or Analysis.