Fix bug with json decoder
Fixes a bug with JSON deserialization of custom user classes defined in the __main__ scope of python scripts and notebooks.

Adds JSON unittests for testing classes, methods, and functions defined in the __main__ module. The previous tests which aimed to do this were not working correctly since their module was being serialized as test.path_to_test_file instead of __main__.