Update CompositeAnalysis initialization (#633)
* Update CompositeAnalysis initialization

Update CompositeAnalysis to be initialized with a list of analysis class objects. This removes the dependence on having the full composite experiment stored in the experiment data to perform composite analysis and makes CompositeAnalysis more natural to subclass for specific cases of component experiments.

* Restructure CompositeAnalysis._run_analysis

Adds a separate helper function to the class to return the list of component experiment data containing the marginalized data. This can be used by subclasses if necessary to obtain marginalized data containers without running analysis.

* Add additional tests