Pin black version
This commit sets a pin on the black version we use in CI and in local
tox jobs. This has historically been done because black was "beta"
software and didn't provide any guarantees on formatting stability
between releases. However, with the recent 22.1.0 release this has
changed and black now guarantees that all formatting will be stable on a
given major version (which is just the year)[1]. This commit pins
experiments to use black 22.x.y to ensure that we don't get spurious
formatting changes moving forward until we're ready to take on a new
major version of black. This will ensure that all contributors will have
consistent formatting locally and in CI.

[1] https://black.readthedocs.io/en/latest/the_black_code_style/index.html#stability-policy