Add json serialization support for data processor and nodes (#639)
* add json serialization support for data processor and nodes

This commit adds __json_encode__ and __json_decode__ methods to processor and node to make them serializable. Data structure of `TrainableDataAction` has also been updated so that it can be stored with trained parameters. RoundTrip tests also added to existing test classes. Note that these objects are compared with repr.

Co-authored-by: Daniel J. Egger <38065505+eggerdj@users.noreply.github.com>