Deprecate FitVal and support saving/loading UFloat (#564)
This PR introduces uncertainties package in data analysis chain in the curve analysis and also replaces FitVal with ufloat value that the package provides. The result figure of curve analysis is now updated to have confidence interval plot with 1 and 3 sigma intervals by defaults. The old database entries with FitVal can be still loaded. Once the entry is loaded, the analysis result value is implicitly typecasted into ufloat with FutureWarning. 

Co-authored-by: Daniel J. Egger <38065505+eggerdj@users.noreply.github.com>
Co-authored-by: Christopher Wood <cjwood@us.ibm.com>