Fix tomography fitter data (#730)
* Fix tomography fitter data to include zero counts

This fixes a bug with tomography fitter data that only included basis elements with non-zero observed counts. This could lead to lower quality fits when used with gaussian weights, and when using constraints for CVXPY and Scipy lstsq fitters.

* Add default TP and trace constraints for cvxpy tomo fitter

The default options for the cvxpy QPT fitters were not including the trace preserving constraint.