black + pylint passs. Have problem with convergence in test_ramsey_xy
The problem in the test "test_ramsey_xy.py" is that the curve fitting function doesn't converge. Looking into the data I see that while the values are different for the counts, the statistics is the same. I think it's because in the new MockIQBackend  I use multinomial distribution function instead of Bernoulli's.