Fix tomography analysis for asymmetric QPT
This fixes a bug with process tomography performed on a different number of input and output qubits where the target channel and TP condition was not calculated correctly when the dimensions of the input and output fitted channel are different.

Also cleans up some TomographyAnalysis helper functions for postprocessing fitter data.