Set number of Aer threads to 1 for macOS test jobs (#780)
In CI on macOS jobs we're seeing occasional job timeouts on the
tomography tests for >= Python 3.8 (which runs serially because
multiprocessing is disabled on macOS by default). These tests spend the
most amount of time in Aer running simulation. One possibility as to
why the jobs are running so much slower is that the small CI nodes are
being overloaded with work by the parallel execution of tests and the
parallel execution of Aer. This commit updates the tox config to pass
through the env variable to set the number of OpenMP threads and then
updates the macOS job configuration to set the number of threads to 1 to
rule this out. While this may not fix the issue it'll be useful in
debugging the cause of the timeouts.

Co-authored-by: Christopher J. Wood <cjwood@us.ibm.com>