Test newer checkout action in docs job
In the recent 0.3.0 release we noticed that the way the checkout action
works has been prevented reno from working correctly. The way the
checkout action handles checking out a ref isn't quite inside the
expectations of reno (and honestly a bit bizzare IMO) and it is
preventing the history scan that reno does to find releases and which
note files are part of each release from finding the full history. To
attempt to fix this issue this commit switches the version of the
checkout action to the latest version. Hopefully this will change the
git usage of the action and fix how

If this doesn't work we can either try using git directly in the docs
jobs to force it to do a normal checkout and preserve the history or
update the reno directive usage to explicitly list all the versions and
force reno to manually scan all the stable branches.