Set earliest version in reno config
This commit sets the earliest_version configuration option for reno [1]
which is used to specify the earliest version to include. In local
testing with how the github actions checkout action works without this
set the reno history scan stops processing requests at the first tag it
encounters but by setting this option it scans the entire history.

[1] https://docs.openstack.org/reno/latest/user/usage.html#configuring-reno