Mark support for for Python 3.10 and add test jobs (#810)
This commit adds a missing trove classifier from the package metadata
indicating that qiskit-experiments supports and works on Python 3.10.
This was missing before even though experiments supports 3.10 without
issue. Additionally this commit adds the missing CI configuration for
running tests on 3.10. This way we're testing that we don't regress and
break 3.10 support while it's a supported version.