Update tox config to avoid sdist for tox 4.x.x compat
This commit updates the tox configuration to remove the sdist
configuration flag which avoids the intermediate sdist creation as part
of tox building the package under test in it's venv. It is replaced by
the isolated_build option which configures tox to use a PEP 518/PEP 517
mechanism for building experiments. This mirrors the default behavior in
tox 4.0.0. By making this change the tox.ini should be compatible with
both tox 3.x.y and 4.x.y.