Fix barrier instructions in tomography experiments
`QuantumCircuit.barrier()` accepts many forms of input, but `tuple` is
not one of them. The prep and measurement indices passed as tuples to
the barrier method were causing qpy serialization/deserialization to
fail. It is possible that there were other unnoticed consequences to the
type of the barrier argument being incorrect.

Closes https://github.com/Qiskit/qiskit-ibm-provider/issues/523