Add mergify config for auto-backport of labelled PRs
This commit adds a config for mergify that enables a mergify pull
request workflow to automatically backport PRs for stable releases.
We're using github's beta merge queue feature right now so the mergify
queue features aren't needed currently. But, mergify has the support for
automatically cherry-picking/backporting commits from `main` to a stable
branch for us, which github doesn't offer an equivalent feature for yet.
So we can leverage a combination of github merge queues and mergify
autobackport to automate most of are merging and backporting. With this
config any PR that's assigned the label "stable backport potential" will
be automatically backported to the current stable branch post-merge.

The one caveat here is that post release from main we'll need to
remember to bump the stable branch name to the new release. For example,
when we release 0.5.0 we'll have to change stable/0.4 to stable/0.5 in
the config file.