Delay scikit-learn import until first use (#1061)
### Summary

This commit pushes the import of scikit-learn lower into the code, so
that it does not occur until first usage.

### Details and comments

Previously, scikit-learn was in a `try` block which allowed for code
that did not use scikit-learn to work fine when it was not installed.
However, sometimes scikit-learn can be installed but have errors (like
in #1050) which were not caught by the `try` block. Further delaying the
import can help in this case. Additionally, scikit-learn is a little bit
of a slow import, so not importing it when it is not needed gives a
little bit of efficiency (maybe; mostly it imports scipy modules but
those might get imported any way by other analysis code).

---------

Co-authored-by: Helena Zhang <hadron.diaspora@gmail.com>