Add bootstrapped fidelity error bars to tomography experiments (#1074)
<!--
⚠️ If you do not respect this template, your pull request will be
closed.
⚠️ Your pull request title should be short detailed and understandable
for all.
⚠️ Also, please add it in the CHANGELOG file under Unreleased section.
⚠️ If your pull request fixes an open issue, please link to the issue.

✅ I have added the tests to cover my changes.
✅ I have updated the documentation accordingly.
✅ I have read the CONTRIBUTING document.
-->

### Summary

Adds an option to `StateTomographyAnalysis` and
`ProcessTomographyAnalysis` to bootstrap error bars on state and process
fidelity analysis results.

### Details and comments

This can be activated by setting the ``target_bootstrap_samples``
analysis option to a value. Note that bootstrapping involves re-running
the full tomography fit on re-samples of tomography outcome data for
each measurement basis and hence the total analysis time will increase
linearly with the number of bootstrap samples.