Updated PR template (#1083)
Because we've set the merge queue to use the PR description as the
commit message to `main`, it's not good to start the PR template with a
commented block that PR makers don't have to remove. The new template
makes the checklist visible at the end and indicates it should be
deleted.