Refactor T2HahnBackend to use qiskit-aer (#1040)
### Summary

This commit refactors `T2HahnBackend` to use qiskit-aer for improved
performance.

### Details and comments

The T2 tests were quite slow, taking around 149 seconds of run time in
testing. `T2HahnBackend` was doing a form of simple Monte Carlo
simulation, stepping through each circuit and flipping the qubit based
on error probabilities. This behavior scales linearly with number of
shots which is not good when running 1000 shots on around 100 circuits
for each test. This commit switches `T2HahnBackend` to use
`AerSimulator` which simulates each circuit only one time to determine
the final output probability. It can then produce as many shot results
as needed quickly by sampling from this distribution.

With the switch to `AerSimulator`, the run time of the T2 tests dropped
from 149 seconds to 12 seconds. The simulation speed up was more
significant than that because the test time is now not limited by
simulation (simulation is around 0.5 seconds now).