Fix IFS setting in doc deploy script (#1110)
### Summary

The 0.5.0 docs publishing job resulted in failure because the openssl
command to decrypt the credentials file for uploading the docs errored.
The IFS (internal field separator) variable was getting set too broadly
at the the level of the entire script instead of just the tag parsing.
This was causing the passing of the secrets for decrypting the
credential files to fail. This commit updates the IFS usage so it's not
affecting the other commands after we split the tag version. This should
fix the deploy job for future releases.

### Details and comments

(cherry picked from commit 8fe22c965a87d377ffcc4b68eb02c2e6754fcad4)