Bump `qiskit-sphinx-theme` version (#1127)
### Summary

Due to a many-hour outage of the `unpkg` CDN that broke our
documentation top menu, the docs theme has switched over to `jsdelivr`.
Our current docs have been redeployed with the updated `1.11.0rc2`
theme, and this PR sets the version requirement to the highest
compatible release instead of fixing the release version. It also
changes the development branch docs workflow name so it's easily
distinguished in the Actions tab.