deploy documentation in qiskit.org/ecosystem (#1096)
Similar to https://github.com/Qiskit/qiskit-aer/pull/1748, see
https://github.com/Qiskit/qiskit.org/issues/3038 for more context

### Summary

The Qiskit Ecosystem lives in qiskit.org/ecosystem and new place for
qiskit related projects that are not technically Qiskit. This PR adds
that destination for documentation deployment.

### Details and comments

For now, a copy of the same docs in ecosystem would allow to have both
q.o/documentation and q.o/ecosystem for testing that everything is fine.
Once that's done, the CDN will redirects to the new location and the
deploy to q.o/documentation can be removed.

---------

Co-authored-by: Helena Zhang <helena.zhang@ibm.com>