Update release notes and documentation (#1147)
### Summary

This PR formats the release notes for 0.5.1 and updates places in the
`README` and `setup.py` that were still configured for Terra instead of
Experiments, in addition to some formatting fixes and adding
`RoughEFFrequencyCal` to the documentation. There will be another PR to
the 0.5 branch to make the 0.5.1 release notes and bump the patch
version.

(cherry picked from commit 3e09dc2dda9681ee10b3911f231f5a8cc83eb9b3)