Sort process tomography clbits before marginalization
The key behavioral difference for the marginal_distribution() function
is that the input order of the bits of interest is significant and the
bits are passed in out of order this will permute the output from the
marginalization function. In the process tomography case the bits of
interest aren't always in order and if they're out of order the output
doesn't match the expected format. To correct this prior to calling
marginal_distribution() this commit sorts the bits of interest to ensure
the output order is as expected by the fitter.